task authorize user

+++ src/db/managers/task_manager.py
@@ -1,7 +1,7 @@
 from typing import List
 from fastapi import Depends
 from sqlalchemy import func, select, literal_column, String, literal
-from sqlalchemy.sql import label, case, any_, desc, cast
+from sqlalchemy.sql import label, case, any_, desc, cast, or_
 from sqlalchemy.orm import Session
 from sqlalchemy.exc import DatabaseError
 
@@ -28,48 +28,48 @@ class TaskManager:
             raise error
         return task
 
-    def all(self) -> List[Task]:
+    def all(self, user_id: str) -> List[Task]:
         try:
-            tasks = self.db.query(Task).all()
+            tasks = self.db.query(Task).filter((Task.is_visible == True) | (Task.author_id == user_id)).all()
         except DatabaseError as error:
             raise error
         return tasks
 
-    def byId(self, task_id: int):
+    def byId(self, user_id: str, task_id: int):
         try:
-            answer = self.db.query(Task).filter(Task.task_id == task_id).one()
+            answer = self.db.query(Task).filter((Task.is_visible == True) | (Task.author_id == user_id)).filter(Task.task_id == task_id).one()
         except DatabaseError as error:
             raise error
         return answer
 
-    def delete(self, task:TaskBase):
+    def delete(self, user_id: str, task:TaskBase):
         try:
-            self.db.query(Task).filter(Task.task_id == task.task_id).delete()
+            self.db.query(Task).filter(Task.task_id == task.task_id).filter(Task.author_id == user_id).delete()
             self.db.commit()
         except DatabaseError as error:
             self.db.rollback()
             raise error
         return
 
-    def update(self, task:TaskModel):
+    def update(self, user_id: str, task:TaskModel):
         try:
-            self.db.query(Task).filter(Task.task_id == task.task_id).update(task.dict())
+            self.db.query(Task).filter(Task.task_id == task.task_id).filter(Task.author_id == user_id).update(task.dict())
             self.db.commit()
             db_task = self.db.query(Task).filter(Task.task_id == task.task_id).one()
         except DatabaseError as error:
             raise error
         return db_task
 
-    def one_with_answers(self, task_id:int):
+    def one_with_answers(self, user_id: str, task_id: int):
         try:
-            task = self.db.query(Task).filter(Task.task_id == task_id).one()
+            task = self.db.query(Task).filter(Task.task_id == task_id).filter((Task.is_visible == True) | (Task.author_id == user_id)).one()
         except DatabaseError as error:
             raise error
         return task
 
-    def all_with_answers(self):
+    def all_with_answers(self, user_id: str):
         try:
-            tasks = self.db.query(Task).all()
+            tasks = self.db.query(Task).filter((Task.is_visible == True) | (Task.author_id == user_id)).all()
         except DatabaseError as error:
             raise error
         return tasks
@@ -89,9 +89,10 @@ class TaskManager:
             raise error
         return db_task
     
-    def find(self, tags: List[int], search_string: str = None, subject_code: str = None, offset: int = 0, limit: int = 25):
+    def find(self, user_id: str, tags: List[int], search_string: str = None, subject_code: str = None, offset: int = 0, limit: int = 25):
         try:
-            query = self.db.query(Task)
+            query = self.db.query(Task).\
+                filter((Task.is_visible == True) | (Task.author_id == user_id))
             if len(tags) > 0:
                 query = query.\
                     join(Task.tag_affs).\
